refactor(smtp): extract session_core (state+finish_data+handle_command_line)
